Extract property matchers.
